json-to-clos: initialize inherited slots too
Initialization of an instance of a CLOS class would only take into account the
direct slots declared in the child class and fail to initialize slots in its
parent classes.

This commit ensures all slots, including inherited slots, are considered when
calling JSON-TO-CLOS with a parameter naming a child class.